Created new ProgressBar.jsx component for optimizing purposes and added it to ModalWarning.jsx
